Made steps 32 bit since walking 65535 steps a day is possible

+++ movement.h
@@ -419,7 +419,7 @@ bool movement_enable_step_count(void);
 bool movement_disable_step_count(void);
 bool movement_step_count_is_enabled(void);
 void movement_reset_step_count(void);
-uint16_t movement_get_step_count(void);
+uint32_t movement_get_step_count(void);
 
 // If the board has a temperature sensor, this function will give you the temperature in degrees celsius.
 // If the board has multiple temperature sensors, it will use the most accurate one available.
